backend/requirements.txt
# SWCD Panel System - Python Dependencies

# FastAPI & Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0

# LLM & Embeddings
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-core>=0.1.0
sentence-transformers>=2.3.0

# AWS
boto3>=1.34.0
botocore>=1.34.0

# Environment
python-dotenv>=1.0.0

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Optional: Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0

# Fuzzy matching
rapidfuzz>=3.5.2

